feat(next, vital-demo-next, __vital-next__, __minimal-next__): Add support to run a NextJS site (#2073)
* Add Package @bodiless/next

* Add Next support to @bodiless/vital-image and @bodiless/vital-template

* Add nextJs minimal starter and nextJs vital starter

* Fix static loaded placeholder and allow to define a doc path

* bump up nextjs version

* Add webpack loader which renames all the versions of asTokenSpec in vital packages to direct call to asTokenSpec

* add useLayoutEffect only on client

* Remove DM-sans helmet link from _app.tsx

* Refactor NodeAPi to be in the right package and avoid code duplications

* Fix Component server generated on edit env

* Fix error in case of siteUrl not configured

* package-lock update

* Increase stack-size to prevent error on edit

* Remove vital-test from __vitalnext__

* Apply new useDocsButton and useEditButton to @bodiless/next

* rename sites __vitalnext__ to __vital-next__ and __minimalnext__ to __minimal-next__

* Refactor generated file serving to serve all the file from generated directory

* Add the ability by packages to modify webpack config

* add DM Sans font to __vital-next__

* Rename base tokens with suffix Base

* Add Vital Demo NextJS

* Fix Statoscope plugin and vital-demo-next config files

* Fix dependencies

* Add missing dependency path-browserify

* Fix missing loader dependencies

* Fix null-loader version

* Add package transpilation

* Fix dependencies on @bodiless/backend

* Rename vital and minimal sites for next using only _ in name

* Rename vital-test to vital-demo and remove dependency by __vital__

* Fix lunr path on vital-demo-next

* Create __vital_next__ package

* Address PR feedback, fix default viewport, add ability to pass at site level component befor and after HEAD tags and Body content

* Add GTM tag manager to vital-demo-next

* Fix meta viewport, Image Background placeholder, GitInfo log once

* Fix search to support nextjs generated html files

* docs(bodiless-documentation): Add Curriculum section

- Add top-level "Curriculum" section.
- Add template for lesson pages.

* fix style inclusion

* Make common-tags tree sheakable on next site

* Fix Vital Element Metas to prevent it disappearing when token is shadowed

* Fix Page loading on edit environment

* Refactoring withMeta to modify the original component

* Restore DM-Sans font for vital-demo gatsby

---------

+++ packages/bodiless-navigation/src/Breadcrumbs/asBreadcrumb.tsx
@@ -100,13 +100,14 @@ const asBreadcrumb = ({
       // a layout effect is executed.
       store.setItem(item);
     }
-
-    useLayoutEffect(() => {
-      store.setItem(item);
-      return () => {
-        store.deleteItem(item);
-      };
-    }, [titleNode.data, linkNode.data]);
+    if (typeof window !== 'undefined') {
+      useLayoutEffect(() => {
+        store.setItem(item);
+        return () => {
+          store.deleteItem(item);
+        };
+      }, [titleNode.data, linkNode.data]);
+    }
 
     return (
       <BreadcrumbContextProvider value={item}>
